requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.4.0
discord-webhook>=1.3.0